Allow concurrent live `Session`s
The existing bindings require a `&mut` reference to the context,
indicating that only one session may be alive at a time per context.
However, GlobalPlatform v1.0 says:

However, it is valid for the Client Application to concurrently use
these functions to initialize or finalize different objects; in the
above example two threads could initialize different TEEC_Session
structures.

This indicates both:
1. It should be possible to have two simultaneous live sessions.
2. If synchronization is necessary on the `TEE_Context` object, it is
   the responsibility of the C library to provide that synchronization.

This means that we do not require exclusive access to the `TEE_Context`,
either at creation time or during use of the session, so we can take a
`&` instead.