Add OwnedSession to support Send
By default, Sessions contain a borrow into their Context. This is
correct, because the underlying implementation is allowed to have
pointer pointing back at their context.

However, this makes them harder to use from multiple threads, as a
non-static reference is inherently not Send.

This provides `OwnedSession`, behind the `owned` feature, which keeps a
reference counted hold on the context backing the session to prevent it
from being moved or dropped until the session leaves scope. This way,
the `OwnedSession` can be transferred to another thread which does not
know how long the original thread would keep the context around.